Refined details about the _templates element.
Moved explanations from section 3.2.2 (template keys) to the general templates collection section, as `key` is not an actual field in the nested document. Relaxed the requirement for a template named default to SHOULD and to use the first element of the collection as default template if none explicitly named `default` is present.